// api/api.ts
export interface RegisterIn {
    /**
     * 
     * @type {string}
     * @memberof RegisterIn
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof RegisterIn
     */
    'password': string;
    /**
     * 
     * @type {string}
     * @memberof RegisterIn
     */
    'password_verify': string;
    /**
     * 
     * @type {string}
     * @memberof RegisterIn
     */
    'first_name': string;
    /**
     * 
     * @type {string}
     * @memberof RegisterIn
     */
    'last_name': string;
}
/**
 * 
 * @export
 * @interface TokenObtainPair
 */
export interface TokenObtainPair {
    /**
     * 
     * @type {string}
     * @memberof TokenObtainPair
     */
    'password': string;
    /**
     * 
     * @type {string}
     * @memberof TokenObtainPair
     */
    'email': string;
}
// ...
export interface TokenRefreshPair {
    /**
     * 
     * @type {string}
     * @memberof TokenRefreshPair
     */
    'refresh': string;
}
/**
 * 
 * @export
 * @interface UserSchema
 */
export interface UserSchema {
    /**
     * 
     * @type {string}
     * @memberof UserSchema
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof UserSchema
     */
    'first_name': string;
    /**
     * 
     * @type {string}
     * @memberof UserSchema
     */
    'last_name': string;
}
// ...
export class UserApi extends BaseAPI {
    /**
     * 
     * @summary Me
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public trackerApiViewUserMe(options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).trackerApiViewUserMe(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary New Token
     * @param {TokenObtainPair} tokenObtainPair 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public trackerApiViewUserNewToken(tokenObtainPair: TokenObtainPair, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).trackerApiViewUserNewToken(tokenObtainPair, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Refresh Token
     * @param {TokenRefreshPair} tokenRefreshPair 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public trackerApiViewUserRefreshToken(tokenRefreshPair: TokenRefreshPair, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).trackerApiViewUserRefreshToken(tokenRefreshPair, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Reg Enabled
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public trackerApiViewUserRegEnabled(options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).trackerApiViewUserRegEnabled(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Register
     * @param {RegisterIn} registerIn 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public trackerApiViewUserRegister(registerIn: RegisterIn, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).trackerApiViewUserRegister(registerIn, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Update Me
     * @param {UserSchema} userSchema 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public trackerApiViewUserUpdateMe(userSchema: UserSchema, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).trackerApiViewUserUpdateMe(userSchema, options).then((request) => request(this.axios, this.basePath));
    }
}